Moved exceptions from folder
