+++ src/main/java/ca/utoronto/fitbook/application/port/in/ExerciseInListNotFoundException.java
@@ -1,4 +1,4 @@
-package ca.utoronto.fitbook.application.port.in.exception;
+package ca.utoronto.fitbook.application.port.in;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
